# Conflicts:
#	postman/MES.postman_collection.json
#	sql/create-all-tables.sql
#	sql/dev-insert-test-data.sql
#	src/main/java/com/tde/mescloud/api/rest/SampleController.java
#	src/main/java/com/tde/mescloud/model/converter/EquipmentOutputConverterImpl.java
#	src/main/java/com/tde/mescloud/model/dto/RequestSampleDto.java
#	src/main/java/com/tde/mescloud/model/entity/SampleEntity.java
#	src/main/java/com/tde/mescloud/repository/ComposedProductionOrderRepository.java
#	src/main/java/com/tde/mescloud/repository/HitRepository.java
#	src/main/java/com/tde/mescloud/repository/ProductionOrderRepository.java
#	src/main/java/com/tde/mescloud/repository/ProductionOrderRepositoryImpl.java
#	src/main/java/com/tde/mescloud/service/ComposedProductionOrderService.java
#	src/main/java/com/tde/mescloud/service/ComposedProductionOrderServiceImpl.java
#	src/main/java/com/tde/mescloud/service/FactoryServiceImpl.java
#	src/main/java/com/tde/mescloud/service/HitServiceImpl.java
#	src/main/java/com/tde/mescloud/service/ProductionOrderService.java
#	src/main/java/com/tde/mescloud/service/ProductionOrderServiceImpl.java
#	src/main/java/com/tde/mescloud/service/SampleServiceImpl.java

+++ src/main/java/com/tde/mescloud/model/dto/SampleDto.java
@@ -9,6 +9,6 @@ public class SampleDto {
 
     private Long id;
     private int amount;
-    private double tcaAverage;
-    private double reliability;
+    private int tcaAverage;
+    private int reliability;
 }
